Add support for filtering pakcages (#443)
The amount of packages and their size is getting larger.
We need to have the ability to filter the pakcages that we
whould like to download. Here is the new help output:

```
> ./scripts/local-universe.py --help
usage: local-universe.py [-h] --repository REPOSITORY [--sudo] --out-file
                         OUTFILE [--include INCLUDE]

This script is able to download the latest artifacts for all of the packages
in the Universe repository into a zipfile. It uses a temporary file to store
all of the artifacts as it downloads them because of this it requires that
your temporary filesystem has enough space to store all of the artifact. You
can control the path to the temporary file by setting the TMPDIR environment
variable. E.g. TMPDIR='.' ./scripts/local-universe.py ...

optional arguments:
  -h, --help            show this help message and exit
  --repository REPOSITORY
                        Path to the top level package directory. E.g.
                        repo/packages
  --sudo                Set this if sudo is required when executing the Docker
                        CLI.
  --out-file OUTFILE    Path to the zipfile to use to store all of the
                        resources
  --include INCLUDE     Command separated list of packages to include. If this
                        option is not specified then all packages are
                        downloaded. E.g. --include="marathon,chronos"
```