Updating local-universe.py to use a docker container. (#452)
This changes the local-universe distribution method from a tarball to a self-hosting docker container. The base container (`multi`) contains both nginx and the docker registry. Registry assets are laid out on disk via. a local registry and volume. Both those assets and the http ones (images, non-docker packages) are copied into the final container `mesosphere/universe`.

The CLI interfaces have also changed slightly. `sudo` is now a requirement, `--selected` has been introduced to pull in *only* the packages that have been selected and `jsonschema` is now a requirement for running the script.